pr fixes

+++ api/src/main/kotlin/com/cogoport/ares/api/payment/entity/OrgOutstanding.kt
@@ -1,6 +1,7 @@
 package com.cogoport.ares.api.payment.entity
 
 import io.micronaut.data.annotation.MappedEntity
+import java.math.BigDecimal
 
 @MappedEntity
 data class OrgOutstanding(
@@ -8,9 +9,9 @@ data class OrgOutstanding(
     val organizationName: String?,
     val currency: String?,
     val openInvoicesCount: Int?,
-    val openInvoicesAmount: Float?,
+    val openInvoicesAmount: BigDecimal?,
     val paymentsCount: Int?,
-    val paymentsAmount: Float?,
-    val outstandingAmount: Float?,
+    val paymentsAmount: BigDecimal?,
+    val outstandingAmount: BigDecimal?,
     val zoneCode: String?
 )
