+++ model/src/main/kotlin/com/cogoport/ares/model/payment/DsoResponse.kt
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.annotation.JsonAutoDetect
 import io.micronaut.core.annotation.Introspected
 import io.micronaut.core.annotation.ReflectiveAccess
+import java.math.BigDecimal
 
 @Introspected
 @ReflectiveAccess
@@ -12,5 +13,5 @@ data class DsoResponse(
     @JsonProperty("month")
     val month: String,
     @JsonProperty("dsoForTheMonth")
-    val dsoForTheMonth: Float
+    val dsoForTheMonth: BigDecimal
 )
